+++ src/components/Auth/stories.js
@@ -3,4 +3,5 @@ import { storiesOf } from '@storybook/react';
 import Auth from './Auth';
 
 storiesOf('Auth', module)
-  .add('no props', () => <Auth />);
+  .add('no props', () => <Auth />)
+  .add('currentUserId: 2', () => <Auth currentUserId={2}/>);