+++ src/redux/inputArea.js
@@ -2,6 +2,7 @@ import { Map } from 'immutable';
 import { MAX_MESSAGE_LENGTH } from '../constants';
 import { createSelector } from 'reselect';
 import { types as messagesTypes } from './messages';
+import { types as authTypes } from './auth';
 
 // Types
 
@@ -30,6 +31,8 @@ export default function reducer(state = INITIAL_STATE, action = {}) {
       return state.set('status', 'saving');
     case messagesTypes.SAVE_SUCCEEDED:
       return state.merge({status: 'initial', text: ''});
+    case authTypes.SET_CURRENT_USER_ID:
+      return INITIAL_STATE;
     default: return state;
   }
 }