+++ CHANGELOG.md
@@ -17,6 +17,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Security
 ---
 
+## [0.0.2] 2018-09-07
+
+### Added
+- Fake users and authentication
+- Welcome message
+- Not allow saving empty messages
+
 ## [0.0.12] 2018-09-06
 
 ### Added